refactor: Car enum 필드를 이름, 이동경로, 이동횟수로 결정
- getter & setter 메소드 완성
