+++ src/main/java/racingcar/util/Parsing.java
@@ -1,5 +1,8 @@
 package racingcar;
 
+import racingcar.constant.Message;
+import racingcar.constant.SymbolType;
+
 import java.util.ArrayList;
 import java.util.List;
 import java.util.StringTokenizer;
@@ -8,7 +11,7 @@ public class Parsing {
 
     public List<Car> StringToCarList(String input) {
         List<Car> carList = new ArrayList<>();
-        StringTokenizer carTokenizer = new StringTokenizer(input, Message.DELIMITER_SPLIT_CARS.getValue());
+        StringTokenizer carTokenizer = new StringTokenizer(input, SymbolType.DELIMITER_SPLIT_CARS.getValue());
 
         while (carTokenizer.hasMoreTokens()){
             String temp = carTokenizer.nextToken();
